test: consolidate no-redeclare coverage
